Added a bunch of breaking changes
