+++ gradle-plugins/build.gradle.kts
@@ -48,8 +48,8 @@ group = "tz.co.asoft"
 version = versions.builders
 
 object versions {
-    val kotlin = "1.4.20"
-    val builders = "1.1.1"
+    val kotlin = "1.4.21"
+    val builders = "1.2.0"
     val android_build_tools = "4.1.0"
     val nexus_staging = "0.22.0"
 }
