+++ gradle-plugins/src/main/kotlin/PluginDependenciesSpec.ktx.kt
@@ -1,0 +1,3 @@
+import org.gradle.plugin.use.PluginDependenciesSpec
+
+fun PluginDependenciesSpec.asoft(plugin: String) = id("tz.co.asoft.$plugin")
