+++ gradle-plugins/src/main/kotlin/dependencies.kt
@@ -1,3 +1,4 @@
 import org.jetbrains.kotlin.gradle.plugin.KotlinDependencyHandler
 
 fun KotlinDependencyHandler.asoft(module: String, version: String) = "tz.co.asoft:$module:$version"
+fun KotlinDependencyHandler.kotlinx(module: String, version: String) = "org.jetbrains.kotlinx:kotlinx-$module:$version"
